requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
gtts==2.4.0
SpeechRecognition==3.10.0
requests==2.31.0
sentence-transformers==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
huggingface_hub==0.16.4
markdown2==2.4.10